* Move constellationInit into authWrapper * Move app specific code up to sample layer
